Split mostypes tests into multiple files and increase coverage
