+++ tests/mock_mos/roCreate5.mos.xml
@@ -1,0 +1,9 @@
+<mos>
+  <mosID>MOS ID</mosID>
+  <messageID>1000</messageID>
+  <roCreate>
+    <roID>RO ID</roID>
+    <roSlug>RO SLUG</roSlug>
+    <roEdStart />
+  </roCreate>
+</mos>
