+++ tests/mock_xml/invalid.xml
@@ -1,0 +1,1 @@
+foo
