+++ tests/mostypes/test_mostype_detection.py
@@ -8,10 +8,20 @@ from mosromgr.mostypes import *
 from mosromgr.exc import *
 
 
-def test_mosfile_detect_invalid_xml():
-    "Test we can catch XML parsing errors"
+def test_mosfile_detect_bad_init(rocreate):
+    "Test we can catch consruction using __init__ not _from classmethods"
+    with pytest.raises(TypeError):
+        MosFile(rocreate)
+
+def test_mosfile_detect_invalid_xml_from_file(invalid_xml):
+    "Test we can catch XML parsing errors when constructing from file"
     with pytest.raises(MosInvalidXML):
-        not_xml = "hello"
+        MosFile.from_file(invalid_xml)
+
+def test_mosfile_detect_invalid_xml_from_string():
+    "Test we can catch XML parsing errors when constructing from string"
+    with pytest.raises(MosInvalidXML):
+        not_xml = "foo"
         MosFile.from_string(not_xml)
 
 def test_mosfile_detect_rocreate_contents(rocreate):
@@ -21,7 +31,7 @@ def test_mosfile_detect_rocreate_contents(rocreate):
     """
     xml = rocreate.read_text()
     rc = MosFile.from_string(xml)
-    assert isinstance(rc, RunningOrder)
+    assert type(rc) == RunningOrder
 
 def test_mosfile_detect_rocreate(rocreate):
     """
@@ -29,7 +39,7 @@ def test_mosfile_detect_rocreate(rocreate):
     EXPECT: An object of type RunningOrder
     """
     rc = MosFile.from_file(rocreate)
-    assert isinstance(rc, RunningOrder)
+    assert type(rc) == RunningOrder
 
 def test_mosfile_detect_storysend(rostorysend1):
     """
@@ -37,79 +47,79 @@ def test_mosfile_detect_storysend(rostorysend1):
     EXPECT: An object of type StorySend
     """
     ss = MosFile.from_file(rostorysend1)
-    assert isinstance(ss, StorySend)
+    assert type(ss) == StorySend
 
 def test_mosfile_detect_appendstory(rostoryappend):
     """
-    GIVEN: A path to a appendStory MOS file
+    GIVEN: A path to a roStoryAppend MOS file
     EXPECT: An object of type StoryAppend
     """
     sa = MosFile.from_file(rostoryappend)
-    assert isinstance(sa, StoryAppend)
+    assert type(sa) == StoryAppend
 
 def test_mosfile_detect_storydelete(rostorydelete):
     """
-    GIVEN: A path to a storydelete MOS file
+    GIVEN: A path to a roStoryDelete MOS file
     EXPECT: An object of type StoryDelete
     """
     sd = MosFile.from_file(rostorydelete)
-    assert isinstance(sd, StoryDelete)
+    assert type(sd) == StoryDelete
 
 def test_mosfile_detect_storyinsert(rostoryinsert):
     """
-    GIVEN: A path to a storyinsert MOS file
+    GIVEN: A path to a RoStoryInsert MOS file
     EXPECT: An object of type StoryInsert
     """
     si = MosFile.from_file(rostoryinsert)
-    assert isinstance(si, StoryInsert)
+    assert type(si) == StoryInsert
 
 def test_mosfile_detect_storymove(rostorymove):
     """
-    GIVEN: A path to a storymove MOS file
+    GIVEN: A path to a roStoryMove MOS file
     EXPECT: An object of type StoryMove
     """
     sm = MosFile.from_file(rostorymove)
-    assert isinstance(sm, StoryMove)
+    assert type(sm) == StoryMove
 
 def test_mosfile_detect_storyreplace(rostoryreplace):
     """
-    GIVEN: A path to a storyreplace MOS file
+    GIVEN: A path to a roStoryReplace MOS file
     EXPECT: An object of type StoryReplace
     """
     sr = MosFile.from_file(rostoryreplace)
-    assert isinstance(sr, StoryReplace)
+    assert type(sr) == StoryReplace
 
 def test_mosfile_detect_itemdelete(roitemdelete):
     """
-    GIVEN: A path to a itemdelete MOS file
+    GIVEN: A path to a roItemDelete MOS file
     EXPECT: An object of type ItemDelete
     """
     id = MosFile.from_file(roitemdelete)
-    assert isinstance(id, ItemDelete)
+    assert type(id) == ItemDelete
 
 def test_mosfile_detect_iteminsert(roiteminsert):
     """
-    GIVEN: A path to a iteminsert MOS file
+    GIVEN: A path to a roItemInsert MOS file
     EXPECT: An object of type ItemInsert
     """
     ii = MosFile.from_file(roiteminsert)
-    assert isinstance(ii, ItemInsert)
+    assert type(ii) == ItemInsert
 
 def test_mosfile_detect_movemultipleitem(roitemmovemultiple):
     """
-    GIVEN: A path to a itemMoveMultiple MOS file
+    GIVEN: A path to a roItemMoveMultiple MOS file
     EXPECT: An object of type ItemMoveMultiple
     """
     imm = MosFile.from_file(roitemmovemultiple)
-    assert isinstance(imm, ItemMoveMultiple)
+    assert type(imm) == ItemMoveMultiple
 
 def test_mosfile_detect_replaceitem(roitemreplace):
     """
-    GIVEN: A path to a replaceItem MOS file
+    GIVEN: A path to a roReplaceItem MOS file
     EXPECT: An object of type ItemReplace
     """
     ir = MosFile.from_file(roitemreplace)
-    assert isinstance(ir, ItemReplace)
+    assert type(ir) == ItemReplace
 
 def test_mosfile_detect_roreplace(roreplace):
     """
@@ -117,7 +127,7 @@ def test_mosfile_detect_roreplace(roreplace):
     EXPECT: An object of type RunningOrderReplace
     """
     rr = MosFile.from_file(roreplace)
-    assert isinstance(rr, RunningOrderReplace)
+    assert type(rr) == RunningOrderReplace
 
 def test_mosfile_detect_elementaction_replace_story(roelementactionstoryreplace):
     """
@@ -126,7 +136,7 @@ def test_mosfile_detect_elementaction_replace_story(roelementactionstoryreplace)
     EXPECT: An object of type EAStoryReplace
     """
     ea = MosFile.from_file(roelementactionstoryreplace)
-    assert isinstance(ea, EAStoryReplace)
+    assert type(ea) == EAStoryReplace
 
 def test_mosfile_detect_elementaction_replace_item(roelementactionitemreplace):
     """
@@ -135,7 +145,7 @@ def test_mosfile_detect_elementaction_replace_item(roelementactionitemreplace):
     EXPECT: An object of type EAItemReplace
     """
     ea = MosFile.from_file(roelementactionitemreplace)
-    assert isinstance(ea, EAItemReplace)
+    assert type(ea) == EAItemReplace
 
 def test_mosfile_detect_elementaction_delete_story(roelementactionstorydelete):
     """
@@ -144,7 +154,7 @@ def test_mosfile_detect_elementaction_delete_story(roelementactionstorydelete):
     EXPECT: An object of type EAStoryDelete
     """
     ea = MosFile.from_file(roelementactionstorydelete)
-    assert isinstance(ea, EAStoryDelete)
+    assert type(ea) == EAStoryDelete
 
 def test_mosfile_detect_elementaction_delete_item(roelementactionitemdelete):
     """
@@ -153,7 +163,7 @@ def test_mosfile_detect_elementaction_delete_item(roelementactionitemdelete):
     EXPECT: An object of type EAItemDelete
     """
     ea = MosFile.from_file(roelementactionitemdelete)
-    assert isinstance(ea, EAItemDelete)
+    assert type(ea) == EAItemDelete
 
 def test_mosfile_detect_elementaction_insert_story(roelementactionstoryinsert):
     """
@@ -162,7 +172,7 @@ def test_mosfile_detect_elementaction_insert_story(roelementactionstoryinsert):
     EXPECT: An object of type EAStoryInsert
     """
     ea = MosFile.from_file(roelementactionstoryinsert)
-    assert isinstance(ea, EAStoryInsert)
+    assert type(ea) == EAStoryInsert
 
 def test_mosfile_detect_elementaction_insert_item(roelementactioniteminsert):
     """
@@ -171,7 +181,7 @@ def test_mosfile_detect_elementaction_insert_item(roelementactioniteminsert):
     EXPECT: An object of type EAItemInsert
     """
     ea = MosFile.from_file(roelementactioniteminsert)
-    assert isinstance(ea, EAItemInsert)
+    assert type(ea) == EAItemInsert
 
 def test_mosfile_detect_elementaction_swap_story(roelementactionstoryswap):
     """
@@ -180,7 +190,7 @@ def test_mosfile_detect_elementaction_swap_story(roelementactionstoryswap):
     EXPECT: An object of type EAStorySwap
     """
     ea = MosFile.from_file(roelementactionstoryswap)
-    assert isinstance(ea, EAStorySwap)
+    assert type(ea) == EAStorySwap
 
 def test_mosfile_detect_elementaction_swap_item(roelementactionitemswap):
     """
@@ -189,7 +199,7 @@ def test_mosfile_detect_elementaction_swap_item(roelementactionitemswap):
     EXPECT: An object of type EAItemSwap
     """
     ea = MosFile.from_file(roelementactionitemswap)
-    assert isinstance(ea, EAItemSwap)
+    assert type(ea) == EAItemSwap
 
 def test_mosfile_detect_elementaction_move_story(roelementactionstorymove):
     """
@@ -198,7 +208,7 @@ def test_mosfile_detect_elementaction_move_story(roelementactionstorymove):
     EXPECT: An object of type EAStoryMove
     """
     ea = MosFile.from_file(roelementactionstorymove)
-    assert isinstance(ea, EAStoryMove)
+    assert type(ea) == EAStoryMove
 
 def test_mosfile_detect_elementaction_move_item(roelementactionitemmove):
     """
@@ -207,15 +217,15 @@ def test_mosfile_detect_elementaction_move_item(roelementactionitemmove):
     EXPECT: An object of type EAItemMove
     """
     ea = MosFile.from_file(roelementactionitemmove)
-    assert isinstance(ea, EAItemMove)
+    assert type(ea) == EAItemMove
 
 def test_mosfile_detect_metadata_replace(rometadatareplace):
     """
-    GIVEN: A path to a metadataReplace MOS file
+    GIVEN: A path to a roMetadataReplace MOS file
     EXPECT: An object of type MetaDataReplace
     """
     ea = MosFile.from_file(rometadatareplace)
-    assert isinstance(ea, MetaDataReplace)
+    assert type(ea) == MetaDataReplace
 
 def test_mosfile_detect_delete(rodelete):
     """
@@ -223,12 +233,20 @@ def test_mosfile_detect_delete(rodelete):
     EXPECT: An object of type RunningOrderEnd
     """
     rd = MosFile.from_file(rodelete)
-    assert isinstance(rd, RunningOrderEnd)
+    assert type(rd) == RunningOrderEnd
 
 def test_mosfile_detect_readytoair(roreadytoair):
+    """
+    GIVEN: A path to a roReadyToAir MOS file
+    EXPECT: An object of type ReadyToAir
+    """
     rta = MosFile.from_file(roreadytoair)
-    assert isinstance(rta, ReadyToAir)
+    assert type(rta) == ReadyToAir
 
 def test_get_mos_object_invalid_mos_type(roinvalid):
+    """
+    GIVEN: A path to a roDelete MOS file
+    EXPECT: An object of type RunningOrderEnd
+    """
     with pytest.raises(UnknownMosFileType):
         MosFile.from_file(roinvalid)
